Create best time to buy and sell stocks(B)
198/212

makes 2 lists then 2 loops such that inner loop starts from the next index of the outer loop. append the difference into one loop and after the termination of the inner loop save the biggest number from the list into an another list. at end the end of the outer loop completely save the biggest number in a variable from the 2nd list and return it is greater than 0 else return 0.