Product of array self (E)
make 2 list pre and post
pre the cumulative sum from beginning and post has cumulative from the end next
traverse as much of length in the list
multiple the previous value of current index from pre list and next value of current index from post list. in case out of bound then just multiply with 1 in place of the out of index list. use if else to find the out of index situation.